backend/requirements.txt
Flask==3.0.0
Flask-CORS==4.0.0
python-dotenv==1.0.0
numpy==1.26.2
opencv-python==4.8.1.78
Pillow==10.1.0
dlib==19.24.2
firebase-admin==6.3.0
google-cloud-storage==2.13.0
requests==2.31.0
gunicorn==21.2.0
scikit-learn==1.3.2
face-recognition==1.3.0
imutils==0.5.4